# [CR] Dtype inference and control in CQT