--- a/librosa/core/constantq.py
+++ b/librosa/core/constantq.py
@@ -953,10 +953,17 @@ def __trim_stack(cqt_resp, n_bins, dtype):
     cqt_out = np.empty((n_bins, max_col), dtype=dtype, order='F')
 
     # Copy per-octave data into output array
-    n = 0
-    for c_i in cqt_resp[::-1]:
-        cqt_out[n:min(n + c_i.shape[0], n_bins)] = c_i[:min(c_i.shape[0], n_bins - n), :max_col]
-        n += c_i.shape[0]
+    end = n_bins
+    for c_i in cqt_resp:
+        # By default, take the whole octave
+        n_oct = c_i.shape[0]
+        # If the whole octave is more than we can fit,
+        # take the highest bins from c_i
+        if end < n_oct:
+            n_oct = n_oct - end
+
+        cqt_out[end - n_oct:end] = c_i[-n_oct:, :max_col]
+        end -= n_oct
 
     return cqt_out