--- a/tests/test_constantq.py
+++ b/tests/test_constantq.py
@@ -69,7 +69,7 @@ def sr_cqt():
 
 @pytest.fixture(scope="module")
 def y_cqt(sr_cqt):
-    return make_signal(sr_cqt, 5.0)
+    return make_signal(sr_cqt, 2.0)
 
 
 @pytest.mark.xfail(raises=librosa.ParameterError)
@@ -313,7 +313,7 @@ def sr_white():
 @pytest.fixture(scope="module")
 def y_white(sr_white):
     srand()
-    return np.random.randn(30 * sr_white)
+    return np.random.randn(10 * sr_white)
 
 
 @pytest.mark.parametrize("scale", [False, True])