--- a/tests/test_constantq.py
+++ b/tests/test_constantq.py
@@ -520,3 +520,9 @@ def test_griffinlim_cqt_momentum_warn():
     x = np.zeros((33, 3))
     with pytest.warns(UserWarning):
         librosa.griffinlim_cqt(x, momentum=2)
+
+
+@pytest.mark.parametrize('dtype', [np.complex64, np.complex128])
+def test_cqt_precision(y_cqt, sr_cqt, dtype):
+    C = librosa.cqt(y=y_cqt, sr=sr_cqt, dtype=dtype)
+    assert np.dtype(C.dtype) == np.dtype(dtype)