#### Reference Issue
- Fixes #1143 
- Fixes #1144 
- Addresses #1074 


#### What does this implement/fix? Explain your changes.

This PR addresses dtype control in spectrogram calculations in a few steps:

1. Helper functions are added to `util` to infer minimal dtypes when converting between real and complex
2. `sparsify_rows` API has expanded to allow explicit dtype control (this supports CQT basis construction)
3. all CQT-related methods now have a `dtype=` parameter
4. all STFT-related methods (including GL) do automatic dtype inference

#### Any other comments?

Committing this now to get the CI on the way.

As mentioned in #1144, I'd also like to revamp the `__trim_stack` helper in cqt to be more memory efficient.  Rolling that into this PR because this is broadly concerned with memory efficiency and copying.