@bmcfee 2020-06-19
trim_stack rewrite is done, and this is ready for CR.

I wound up not using numba acceleration because there wasn't much benefit, and the complexity of handling typed lists (list of ndarrays) didn't seem worth it.

I get about a 10ms improvement on nutcracker-HQ with 36bpo, but the bigger savings are in memory usage because an entire set of copies of the output are no longer needed.