@bmcfee 2020-06-19
Ok, I think this one's good to go.

Quick benchmark for the new trim_stack:
**Before**
```python
In [3]: y, sr = librosa.load(librosa.ex('nutcracker', hq=True), sr=None)                      

In [4]: %timeit librosa.cqt(y=y, sr=sr, bins_per_octave=36, n_bins=216)                       
2.59 s ± 5.87 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```
**After**
```python
In [5]: %timeit librosa.cqt(y=y, sr=sr, bins_per_octave=36, n_bins=216)                       
2.33 s ± 5.07 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```